+++ test/cache_templates_test.rb
@@ -18,10 +18,11 @@ class CacheTemplatesTest < ActiveSupport::TestCase
 
   test "cached templates should not be modifiable in place" do
     ActionController::Base.stub(:perform_caching).and_return(true)
-    @library.compile_template_from_source('', 'some/path')
-    t = @library.compile_template_from_source("attribute :id", 'some/path')
+    t = @library.compile_template_from_source('', 'some/path')
 
-    assert_equal({}, t.source)
+    t.merge!(:_data => :foo)
+
+    assert_equal({}, @library.compile_template_from_path('some/path').source)
   end
 
   test "don't cache templates cache_templates is enabled but perform_caching is not active" do